fixed rare recomb on basal branch bug

+++ src/thread.h
@@ -52,7 +52,7 @@ void sample_arg_removal_path(LocalTrees *trees, int node, int *path);
 void sample_arg_removal_path(LocalTrees *trees, int node, int pos, int *path);
 void sample_arg_removal_leaf_path(LocalTrees *trees, int node, int *path);
 
-void sample_arg_removal_path_recomb(LocalTrees *trees, float recomb_preference,
+void sample_arg_removal_path_recomb(LocalTrees *trees, double recomb_preference,
                                     int *path);
